Answer more networking questions:
Questions:

What type of network a computer in local is connected to?

 1. Internet
 2. WAN
 3. LAN

What type of network could connect an office in one building to another
office in a building a few streets away?

 1. Internet
 2. WAN
 3. LAN

What network do you use when you browse www.google.com from your
smartphone (not connected to the wifi)?

 1. Internet
 2. WAN
 3. LAN